+++ packages/client/src/tx/signed-extensions/chain/CheckGenesis.ts
@@ -1,8 +1,7 @@
 import { map } from "rxjs"
 import type { GetChainSignedExtension } from "../internal-types"
-import { empty, genesisHashFromCtx } from "../utils"
+import { additionalSigned } from "../utils"
+import { fromHex } from "@polkadot-api/utils"
 
 export const CheckGenesis: GetChainSignedExtension = (ctx) =>
-  genesisHashFromCtx(ctx).pipe(
-    map((additionalSigned) => ({ value: empty, additionalSigned })),
-  )
+  ctx.chainHead.genesis$.pipe(map(fromHex), map(additionalSigned))