+++ packages/client/src/tx/signed-extensions/chain/CheckSpecVersion.ts
@@ -1,8 +1,6 @@
 import { map } from "rxjs"
 import type { GetChainSignedExtension } from "../internal-types"
-import { empty, systemVersionProp$ } from "../utils"
+import { additionalSigned, systemVersionProp$ } from "../utils"
 
 export const CheckSpecVersion: GetChainSignedExtension = ({ lookupFn }) =>
-  systemVersionProp$("spec_version", lookupFn).pipe(
-    map((additionalSigned) => ({ additionalSigned, value: empty })),
-  )
+  systemVersionProp$("spec_version", lookupFn).pipe(map(additionalSigned))